masked noisy approach computation added
Based on Akrame-et.al-2014 we added noise and mask to maps.
we now intend to subtract mean of same patch of sims from all data sets and divide the result
to the variance of the measure (in this case the measure is Var(T) which means
we divide to Var(Var(T))) to find the significance of our results.